move op related types to core/ops.rs
